# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Online Bookstore API",
        "description": "Online Bookstore Management System",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/inventory": {
            "get": {
                "tags": [
                    "Inventory"
                ],
                "summary": "Inventory List",
                "operationId": "c8f96496342baec381c44a880ef4510b",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "book": {
                                                        "properties": {
                                                            "title": {
                                                                "type": "string",
                                                                "example": "Burning Heat"
                                                            },
                                                            "author": {
                                                                "type": "string",
                                                                "example": "Akiyoshi Rikako"
                                                            },
                                                            "publisher": {
                                                                "type": "string",
                                                                "example": "Elexmedia"
                                                            },
                                                            "year_released": {
                                                                "type": "string",
                                                                "example": "2021"
                                                            },
                                                            "genre": {
                                                                "type": "string",
                                                                "example": "thriller"
                                                            },
                                                            "price": {
                                                                "type": "integer",
                                                                "example": "120000"
                                                            }
                                                        },
                                                        "type": "object"
                                                    },
                                                    "Stock": {
                                                        "type": "integer",
                                                        "example": 10
                                                    },
                                                    "is_available": {
                                                        "type": "boolean",
                                                        "example": true
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Inventory"
                ],
                "summary": "Inventory Store",
                "operationId": "92df36cfb81355e2d905b7d71d541f42",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "book_id": {
                                                "type": "integer"
                                            },
                                            "stock": {
                                                "type": "integer"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "book_id": 1,
                                    "stock": 10
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "book": {
                                                    "properties": {
                                                        "title": {
                                                            "type": "string",
                                                            "example": "Burning Heat"
                                                        },
                                                        "author": {
                                                            "type": "string",
                                                            "example": "Akiyoshi Rikako"
                                                        },
                                                        "publisher": {
                                                            "type": "string",
                                                            "example": "Elexmedia"
                                                        },
                                                        "year_released": {
                                                            "type": "string",
                                                            "example": "2021"
                                                        },
                                                        "genre": {
                                                            "type": "string",
                                                            "example": "thriller"
                                                        },
                                                        "price": {
                                                            "type": "integer",
                                                            "example": "120000"
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "Stock": {
                                                    "type": "integer",
                                                    "example": 10
                                                },
                                                "is_available": {
                                                    "type": "boolean",
                                                    "example": true
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
        "/api/inventory/{inventory}": {
            "get": {
                "tags": [
                    "Inventory"
                ],
                "summary": "Inventory Detail",
                "operationId": "a49991e4d228f087c9980a00702e85af",
                "parameters": [
                    {
                        "name": "inventory",
                        "in": "path",
                        "description": "Inventory Id / Unique Key",
                        "required": true,
                        "schema": {
                            "type": "string|integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "book": {
                                            "properties": {
                                                "title": {
                                                    "type": "string",
                                                    "example": "Burning Heat"
                                                },
                                                "author": {
                                                    "type": "string",
                                                    "example": "Akiyoshi Rikako"
                                                },
                                                "publisher": {
                                                    "type": "string",
                                                    "example": "Elexmedia"
                                                },
                                                "year_released": {
                                                    "type": "string",
                                                    "example": "2021"
                                                },
                                                "genre": {
                                                    "type": "string",
                                                    "example": "thriller"
                                                },
                                                "price": {
                                                    "type": "integer",
                                                    "example": "120000"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "Stock": {
                                            "type": "integer",
                                            "example": 10
                                        },
                                        "is_available": {
                                            "type": "boolean",
                                            "example": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Inventory"
                ],
                "summary": "Inventory Delete",
                "operationId": "c75e576932ab2e6f129cdcad8d7d14ac",
                "parameters": [
                    {
                        "name": "inventory",
                        "in": "path",
                        "description": "Inventory Id / Unique Key",
                        "required": true,
                        "schema": {
                            "type": "string|integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "boolean",
                                            "example": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Inventory"
                ],
                "summary": "Inventory Update",
                "operationId": "ee74cfefd0505bc855e0c97123df6818",
                "parameters": [
                    {
                        "name": "inventory",
                        "in": "path",
                        "description": "Inventory Id / Unique Key",
                        "required": true,
                        "schema": {
                            "type": "string|integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "book_id": {
                                                "type": "integer"
                                            },
                                            "stock": {
                                                "type": "integer"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "book_id": 1,
                                    "stock": 10
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "book": {
                                            "properties": {
                                                "title": {
                                                    "type": "string",
                                                    "example": "Burning Heat"
                                                },
                                                "author": {
                                                    "type": "string",
                                                    "example": "Akiyoshi Rikako"
                                                },
                                                "publisher": {
                                                    "type": "string",
                                                    "example": "Elexmedia"
                                                },
                                                "year_released": {
                                                    "type": "string",
                                                    "example": "2021"
                                                },
                                                "genre": {
                                                    "type": "string",
                                                    "example": "thriller"
                                                },
                                                "price": {
                                                    "type": "integer",
                                                    "example": "120000"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "Stock": {
                                            "type": "integer",
                                            "example": 10
                                        },
                                        "is_available": {
                                            "type": "boolean",
                                            "example": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "token": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}